requirements.txt
aiohttp~=3.11.18
Telethon~=1.40.0
python-dotenv==1.0.0
aiohttp_socks
pyqt5
tgcrypto